update-logic-nav-to-editing-if-no-user_info
