--- src/components/layout/AuthSuccess.tsx
@@ -1,10 +1,0 @@
-export default function AuthSuccess() {
-    return (
-      <div className="w-full min-h-screen">
-        <section className="w-full h-screen flex justify-center items-center">
-          <h1 className="text-amber-50">這是登入頁面</h1>
-        </section>
-      </div>
-    );
-  }
-
